+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 > **Note**: odd version numbers, for example, `0.13.0`, are not included in this changelog. They are used to test the new features and fixes before the final release.
 
-## [0.18.0] - 2025-02-27
+## [0.18.1] - 2025-02-27
 
 ### Added:
 
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Diagnostics: Fixed GraphMockResponsePlugin does not require a config section issue
 - Diagnostics: Fixed languageModel property being reported as invalid config section
+- Install: Fixed broken link for Linux
+- Install: Updated brew tap command to reference new tap location
 
 ## [0.16.0] - 2025-02-03